Split ConsoleMessage mapping
