+++ browser/mapping.go
@@ -64,37 +64,6 @@ func parseWaitForFunctionArgs(
 	return js, popts, exportArgs(gargs), nil
 }
 
-// mapConsoleMessage to the JS module.
-func mapConsoleMessage(vu moduleVU, cm *common.ConsoleMessage) mapping {
-	rt := vu.Runtime()
-	return mapping{
-		"args": func() *goja.Object {
-			var (
-				margs []mapping
-				args  = cm.Args
-			)
-			for _, arg := range args {
-				a := mapJSHandle(vu, arg)
-				margs = append(margs, a)
-			}
-
-			return rt.ToValue(margs).ToObject(rt)
-		},
-		// page(), text() and type() are defined as
-		// functions in order to match Playwright's API
-		"page": func() *goja.Object {
-			mp := mapPage(vu, cm.Page)
-			return rt.ToValue(mp).ToObject(rt)
-		},
-		"text": func() *goja.Object {
-			return rt.ToValue(cm.Text).ToObject(rt)
-		},
-		"type": func() *goja.Object {
-			return rt.ToValue(cm.Type).ToObject(rt)
-		},
-	}
-}
-
 // mapBrowser to the JS module.
 func mapBrowser(vu moduleVU) mapping { //nolint:funlen
 	rt := vu.Runtime()
